% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{attributes_l}
\alias{attributes_l}
\title{Attributes lens}
\format{An object of class \code{lens} of length 2.}
\usage{
attributes_l
}
\description{
The lens equivalent of \link{attributes} and \link{attributes<-}
}
\keyword{datasets}
