% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{attrl}
\alias{attrl}
\title{Construct a lens into an attribute}
\usage{
attrl(attrib)
}
\arguments{
\item{attrib}{A length one character vector indicating
the attribute to lens into.}
}
\description{
The lens version of \code{attr} and \code{attr<-}
}
