% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{body_l}
\alias{body_l}
\title{Body lens}
\format{An object of class \code{lens} of length 2.}
\usage{
body_l
}
\description{
A lens into the body of a function. The
lens equivalent of \link{body} and \link{body<-}.
You probably shouldn't use this.
}
\examples{
inc2 <- function(x) x + 2
view(inc2, body_l)
inc4 <- set(inc2, body_l, quote(x + 4))
inc4(10)
}
\keyword{datasets}
