% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-lenses.R
\name{collapse_l}
\alias{collapse_l}
\title{Collapse Lens}
\usage{
collapse_l(collapse)
}
\arguments{
\item{collapse}{The collapse string for \code{view} and the split string for \code{set}.}
}
\description{
\link{view} renders a list of character vectors as a vector of strings. Elements of
the input character vectors are pasted together with \code{paste0(v, collapse = collapse)}.
\link{set} replaces the input character vector with the replacement
data \link{view}ed with \href{"collapse"}{strsplit_l}.
}
\details{
Essentially the opposite of \link{strsplit_l}
}
\examples{
view(list(c("a", "b", "c")), collapse_l("-"))
set(list(c("a", "b", "c")), collapse_l("-"), c("a-b"))
set(list(c("a", "b", "c")), collapse_l("-"), c("qr-pz"))
}
