% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\docType{data}
\name{colnamesl}
\alias{colnamesl}
\title{A lens into the column names of an object}
\format{An object of class \code{lens} of length 2.}
\usage{
colnamesl
}
\description{
The lens version of \code{colnames} and \code{colnames<-}
}
\keyword{datasets}
