% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{cols_l}
\alias{cols_l}
\title{Column lens}
\usage{
cols_l(cols, drop = FALSE)
}
\arguments{
\item{cols}{the columns to focus on}

\item{drop}{whether or not to drop dimensions with length 1}
}
\description{
Create a lens into a set of columns
}
