% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\docType{data}
\name{diag_l}
\alias{diag_l}
\title{Lens into the diagonal of a matrix}
\format{An object of class \code{lens} of length 2.}
\usage{
diag_l
}
\description{
A lens into a matrix's diagonal elements
}
\keyword{datasets}
