% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{dim_l}
\alias{dim_l}
\title{Dims lens}
\format{An object of class \code{lens} of length 2.}
\usage{
dim_l
}
\description{
A lens into an objects dimensions
}
\examples{
x <- 1:10

(y <- set(x, dim_l, c(2,5)))
view(y, dim_l)
}
\keyword{datasets}
