% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{drop_while_il}
\alias{drop_while_il}
\title{Conditional trim lens}
\usage{
drop_while_il(f)
}
\arguments{
\item{f}{the predicate (logical) function}
}
\description{
A lens into all elements starting from the first
element that doesn't satisfy a predicate. Essentially
the complement of \link{take_while_l}
}
