% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{env_l}
\alias{env_l}
\title{Environment lens}
\format{An object of class \code{lens} of length 2.}
\usage{
env_l
}
\description{
A lens into the environment of an object. This
is the lens version of \link{environment} and \link{environment<-}
}
\keyword{datasets}
