% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\docType{data}
\name{envl}
\alias{envl}
\title{Environment lens}
\format{An object of class \code{lens} of length 2.}
\usage{
envl
}
\description{
A lens into the environment of an object. This
is the lens version of \link{environment} and \link{environment<-}
}
\keyword{datasets}
