% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-lenses.R
\name{fext_l}
\alias{fext_l}
\title{Create lens into the file extension}
\usage{
fext_l(extension = c("last", "first"))
}
\arguments{
\item{extension}{last or first indicating what type of extensions to
focus on.}
}
\description{
A lens that focuses on the extension of a filepath
Inspiration/regex pieces from Jim Hester and r-lib's fs package.
The function creates one of two \code{last} which considers everything
after the last dot (\code{.}) an extension, and \code{first} which considers
everything after the first dot in the basename of the file to be
part of the file extension.
}
