% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{filter_il}
\alias{filter_il}
\title{Filter lens}
\usage{
filter_il(...)
}
\arguments{
\item{...}{unquoted NSE filter arguments}
}
\description{
Create an illegal lens into the result of a filter. Arguments
are interpreted with non-standard evaluation as in
\link[dplyr:filter]{dplyr::filter}
}
