% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-lenses.R
\name{filter_l}
\alias{filter_l}
\title{Filter lens}
\usage{
filter_l(...)
}
\arguments{
\item{...}{unquoted NSE filter arguments}
}
\description{
Create a lawful lens into the result of a filter. This
focuses only columns not involved in the filter condition.
}
\examples{
head(view(iris, filter_l(Species == "setosa"))) # Note Species is not seen
head(over(iris, filter_l(Species == "setosa"), function(x) x + 10))
}
