% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{first_l}
\alias{first_l}
\title{A lens into the first element}
\format{An object of class \code{lens} of length 2.}
\usage{
first_l
}
\description{
Lens version of \code{x[[1]]} and \code{x[[1]] <- val}

[[1]: R:[1
[[1]: R:[1
}
\keyword{datasets}
