% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-lenses.R
\docType{data}
\name{fname_l}
\alias{fname_l}
\title{Lens into the name and extension of a path}
\format{An object of class \code{lens} of length 2.}
\usage{
fname_l
}
\description{
\link{view} is \code{basename}, \link{set} replaces everything after
the last file separator. Trailing file separators
are ignored.
}
\keyword{datasets}
