% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\docType{data}
\name{formals_l}
\alias{formals_l}
\title{Formals lens}
\format{An object of class \code{lens} of length 2.}
\usage{
formals_l
}
\description{
A lens equivalent of \link{formals} and \link{formals<-},
allowing you to change the formal arguments of
a function. As with \link{body_l} you probably shouldn't
use this.
}
\keyword{datasets}
