% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-lenses.R
\docType{data}
\name{fpath_l}
\alias{fpath_l}
\title{Lens into the path of the file}
\format{An object of class \code{lens} of length 2.}
\usage{
fpath_l
}
\description{
This lens extracts the path from a string interpretted as a path.
\link{view} is roughly equivalent to \link{basename}, it differs on how it
handles relative paths with no file separators. \code{view("x/", fpath_l)}
returns \code{NA}, \code{basename("x/")} returns \code{.}. This difference is to satisfy
the \code{view-set} law.
\link{set} also special cases replacing with \code{NA}, obliterating any path components.
\code{set("/a/file/path/", fpath_l, NA)} returns \code{"path/"} contrast this with
\code{set("/a/file/path/", fpath_l, "")} which returns \code{"/path/"}.
}
\examples{
set("/a/file/path/", fpath_l, NA)
set("/a/file/path/", fpath_l, "")
view("x", fpath_l)
view("/a/file/path/", fpath_l)
}
\keyword{datasets}
