% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-lenses.R
\name{fstem_l}
\alias{fstem_l}
\title{Lens into the extensionless file}
\usage{
fstem_l(extension = c("last", "first"))
}
\arguments{
\item{extension}{last or first indicating what type of extensions to
focus on.}
}
\description{
A lens that focuses on the file with the extension removed.
The function creates one of two lenses, \code{last} extension which considers everything
after the last dot (\code{.}) an extension, and \code{first} which considers
everything after the first dot in the basename of the file to be
part of the file extension.
}
\examples{
view("hi.tar.gz", fstem_l())
view("hi.tar.gz", fstem_l("first"))
set("hi.tar.gz", fstem_l(), "new")
set("hi.tar.gz", fstem_l("first"),  "new")
set(c("hi.tar.gz", "woah.txt.zip.gz"), fstem_l("first"), c("file1", "file2"))
set(c("hi.tar.gz", "woah.txt.zip.gz"), fstem_l(), c("file1", "file2"))
}
