% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{index}
\alias{index}
\title{Construct a lens into an index/name}
\usage{
index(el)
}
\arguments{
\item{el}{The element the lens should point to
can be an \code{integer} or name.}
}
\description{
This is the lens version of \code{[[}
}
