% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{indexes}
\alias{indexes}
\title{Construct a lens into a subset of an object}
\usage{
indexes(els)
}
\arguments{
\item{els}{a subset vector, can be \code{integer}, \code{character}
of \code{logical}, pointing to one or more elements of the object}
}
\description{
This is the lens version of \code{[}
}
