% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{levels_l}
\alias{levels_l}
\title{Levels lens}
\format{An object of class \code{lens} of length 2.}
\usage{
levels_l
}
\description{
A lens into the levels of an object. Usually
this is factor levels. Lens
equivalent of \link{levels} and \link{levels<-}.
}
\examples{
x <- factor(c("a", "b"))
view(x, levels_l)
set(x, levels_l, c("A", "B"))
}
\keyword{datasets}
