% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{lget}
\alias{lget}
\title{Lget data with a lens}
\usage{
lget(d, l)
}
\arguments{
\item{d}{the data}

\item{l}{the lens}
}
\description{
Lget the subcomponent of the data referred to by a lens.  This function
merely dispatches to the \code{lget} component of the lens.
}
