% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-lenses.R
\name{lower_tri_l}
\alias{lower_tri_l}
\title{Lens into lower diagonal elements}
\usage{
lower_tri_l(diag = FALSE)
}
\arguments{
\item{diag}{whether or not to include the diagonal}
}
\description{
Create a lens into the lower diagonal elements
of a matrix
}
\examples{
(x <- matrix(1:9, ncol = 3))
view(x, lower_tri_l())
view(x, lower_tri_l(diag = TRUE))
set(x, lower_tri_l(), c(100, 200, 300))
}
