% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{map_over}
\alias{map_over}
\title{Map a function over a lens}
\usage{
map_over(d, l, f)
}
\arguments{
\item{d}{the data}

\item{l}{the lens}

\item{f}{the function to use, potentially a \code{~} specified anonymous function.}
}
\description{
Apply the specified function to each element
of the subobject.
}
