% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\docType{data}
\name{names_l}
\alias{names_l}
\title{A lens into the names of an object}
\format{An object of class \code{lens} of length 2.}
\usage{
names_l
}
\description{
The lens versions of \code{names} and \code{names<-}.
}
\keyword{datasets}
