% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{oscope}
\alias{oscope}
\title{Bind data to a lens}
\usage{
oscope(d, l = idl)
}
\arguments{
\item{d}{The data for interest}

\item{l}{The lens to bind the data to. Defaults to the identity lens}
}
\description{
To flatten lens composition, you can prespecify the data
the lens with be applied to by constructing an objectoscope.
These can be integrated easily with normal data pipelines.
}
\examples{
list(a = 5, b = 1:3, c = 8) \%>\%
  oscope()   \%.\%
  index("b") \%.\%
  index(1)   \%>\%
  lset(10)
}
