% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{over_map}
\alias{over_map}
\title{Map a function over a list lens}
\usage{
over_map(d, l, f)
}
\arguments{
\item{d}{the data}

\item{l}{the lens}

\item{f}{the function to use, potentially a \code{~} specified anonymous function.}
}
\description{
Apply the specified function to each element
of the subobject.
}
