% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{over_with}
\alias{over_with}
\title{Map a function over an in scope lens}
\usage{
over_with(d, l, f)
}
\arguments{
\item{d}{the data}

\item{l}{the lens}

\item{f}{the function to use, potentially a \code{~} specified anonymous function.
The function body is quoted, and evaluated with \code{rlang::eval_tidy(..., data = view(d,l))}}
}
\description{
Apply the specified function with named elements of
the viewed data in scope. Similar to \link[dplyr:mutate]{dplyr::mutate}
}
\examples{
iris \%>\% over_with(id_l, ~ Sepal.Length)
}
