% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-lenses.R
\docType{data}
\name{path2list_l}
\alias{path2list_l}
\title{Convert paths to vectors of path components}
\format{An object of class \code{lens} of length 2.}
\usage{
path2list_l
}
\description{
This is strsplit_l(.Platform$file.sep)
}
\keyword{datasets}
