% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{pluck_l}
\alias{pluck_l}
\title{Pluck as a lens}
\usage{
pluck_l(...)
}
\arguments{
\item{...}{index vectors or lenses}
}
\description{
A lens version of \link[purrr:pluck]{purrr::pluck}. Takes
a series element indicators and creates a composite
lens.
}
\details{
\itemize{
\item length one vectors are converted to \link{index_l},
\item larger vectors are converted to \link{indexes_l},
\item lenses are composed as is.
}

See examples for more
}
\examples{
view(iris, pluck_l("Petal.Length", 10:20, 3))
sepal_l <- index("Sepal.Length")
view(iris, pluck_l(sepal_l, id_l, 3))
}
