% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-lenses.R
\name{reshape_l}
\alias{reshape_l}
\title{Lens into a new dimension(s)}
\usage{
reshape_l(dims)
}
\arguments{
\item{dims}{a vector with the new dimensions}
}
\description{
Construct a lens that is a view of the data with
a new set of dimensions. Both \link{view} and \link{set} check
that the new dimensions match the number of elements of
the data.
}
\examples{
x <- 1:9
view(x, reshape_l(c(3,3)))
set(x, reshape_l(c(3,3)) \%.\% diag_l, 100)
}
