% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\docType{data}
\name{rownamesl}
\alias{rownamesl}
\title{A lens into the row names of an object}
\format{An object of class \code{lens} of length 2.}
\usage{
rownamesl
}
\description{
The lens version of \code{rownames} and \code{rownames<-}
}
\keyword{datasets}
