% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{select_l}
\alias{select_l}
\title{Tidyselect elements by name}
\usage{
select_l(...)
}
\arguments{
\item{...}{An expression to be interpretted by \link[tidyselect:vars_select]{tidyselect::vars_select}
which is the same interpretter as \link[dplyr:select]{dplyr::select}}
}
\description{
Create a lens into a named collection to be lgeted or replaced.
Names of the input are not changed. This generalizes \link[dplyr:select]{dplyr::select}
to arbitrary named collections and allows updating.
}
\examples{
lets <- setNames(seq_along(LETTERS), LETTERS)
lset(lets, select_l(G:F, A, B), 1:4) # A and B are 3,4 for a quick check
}
