% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{send}
\alias{send}
\title{Set one lens to the view of another}
\usage{
send(d, l, m)
}
\arguments{
\item{d}{the data}

\item{l}{the lens to view through}

\item{m}{the lens to set into}
}
\description{
Set one lens to the view of another
}
