% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{set}
\alias{set}
\title{Modify data with a lens}
\usage{
set(d, l, x)
}
\arguments{
\item{d}{the data, or an \link{oscope}}

\item{l}{the lens, or in the case of an \code{oscope}, the replacement}

\item{x}{the replacement value, or nothing in the case of an \code{oscope}}
}
\description{
Set the subcomponent of the data referred to by a lens
with a new value. See \link{lens} for details.  Merely dispatches
to the \code{set} component of the lens.
}
