% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{slab_l}
\alias{slab_l}
\title{Slab lens}
\usage{
slab_l(..., drop = FALSE)
}
\arguments{
\item{...}{arguments as they would be passed \code{[} e.g. \code{x[3,5,7]}

[\code{e.g.}x[3,5,7]: R:%60%20e.g.%20%60x[3,5,7}

\item{drop}{whether or not to drop dimensions with length 1. Only
applies to \code{view}}
}
\description{
Create a lens into a chunk of an array (hyperslab). Uses
the same syntactic rules as \code{[}
}
