% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{slot_l}
\alias{slot_l}
\title{Slot lens}
\usage{
slot_l(slot)
}
\arguments{
\item{slot}{the name of the slot}
}
\description{
The lens equivalent of \code{@} and \code{@<-}
for getting and setting S4 object slots.
}
\examples{
new_class <- setClass("new_class", slots = c(x = "numeric"))
(x <- new_class())

view(x, slot_l("x"))
set(x, slot_l("x"), 1:10)
}
