% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-lenses.R
\name{strsplit_l}
\alias{strsplit_l}
\title{Strsplit lens}
\usage{
strsplit_l(split)
}
\arguments{
\item{split}{The string to split at, this cannot be a regex (yet?).}
}
\description{
A lens into a split representation of a string. Here
\link{view} is almost equivalent to \link{strsplit} with \code{fixed = TRUE},
the difference is that empty strings are included after a
trailing seperator. E.g. \code{view("/x/", strsplit_l("/"))} returns
\code{list(c("", "x", ""))} where \code{strsplit("/x/", "/")} returns
\code{list(c("", "x"))}.
\link{set} is equivalent to \link{paste0}\code{(x, collapse = split)} mapped
over each input with \link{vapply}.
}
\details{
Essentially the opposite of \link{collapse_l}
}
