% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-lenses.R
\name{substr_l}
\alias{substr_l}
\title{A lens into a substring}
\usage{
substr_l(first, last)
}
\arguments{
\item{first}{the starting character}

\item{last}{the last character}
}
\description{
Make a lens to focus on a character range in one or more
strings (character vector).
}
