% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{sym_l}
\alias{sym_l}
\title{Lens into a symmetric view of a matrix}
\usage{
sym_l(diag = FALSE, tri = c("lower", "upper"))
}
\arguments{
\item{diag}{whether or not to include the diagonal}

\item{tri}{which triangle to view}
}
\description{
Create a lens into a symmetric view of a matrix
}
