% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-lenses.R
\docType{data}
\name{t_l}
\alias{t_l}
\title{Matrix transpose lens}
\format{An object of class \code{lens} of length 2.}
\usage{
t_l
}
\description{
Lens into the transpose of a matrix
}
\examples{
(x <- matrix(1:4, ncol = 2))
view(x, t_l)
set(x, t_l, matrix(11:14, ncol = 2))
}
\keyword{datasets}
