% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{take_l}
\alias{take_l}
\title{Construct a lens into a prefix of a vector}
\usage{
take_l(n)
}
\arguments{
\item{n}{number of elements to take, or if negative the
number of elements at the end to not take.}
}
\description{
This constructs a lens into the first \code{n} elements
of an object or the if negative indexing is used,
as many as \code{length(x) - n}.
}
