% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{to_l}
\alias{to_l}
\title{Promote a function to a \code{getter} lens}
\usage{
to_l(f)
}
\arguments{
\item{f}{The function to promote.}
}
\description{
Create a \code{getter} lens from a function.
}
\examples{
# This wouldn't make sense as a general legal lens, but fine as a `getter`
sqrt_l <- to_l(sqrt)
iris_root <- index(1) \%.\% index(1) \%.\% sqrt_l

sqrt(iris[[1]][[1]])
iris \%>\% view(iris_root)
tryCatch(iris \%>\% set(iris_root, 2)
       , error = function(e) "See, can't do that")

}
