% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\docType{data}
\name{transpose_l}
\alias{transpose_l}
\title{Lens into a list of rows}
\format{An object of class \code{lens} of length 2.}
\usage{
transpose_l
}
\description{
A lens that creates a list-of-rows view of a \code{data.frame}
}
\keyword{datasets}
