% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{unlist_l}
\alias{unlist_l}
\title{Unlist lens}
\format{An object of class \code{lens} of length 2.}
\usage{
unlist_l
}
\description{
A lens between a list and an unrecursively \link{unlist}ed object.
}
\examples{
(x <- list(x = list(y = 1:10)))
view(x, unlist_l)
set(x, unlist_l \%.\% unlist_l, rep("hello", 10))
}
\keyword{datasets}
