% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{view}
\alias{view}
\title{View data with a lens}
\usage{
view(d, l)
}
\arguments{
\item{d}{the data}

\item{l}{the lens}
}
\description{
Get the subcomponent of the data referred to by a lens.  This function
merely dispatches to the \code{view} component of the lens.
}
