% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\docType{data}
\name{class_l}
\alias{class_l}
\title{Class lens}
\format{An object of class \code{lens} of length 2.}
\usage{
class_l
}
\description{
A lens into the class of an object. Lens
equivalent of \link{class} and \link{class<-}.
}
\keyword{datasets}
