% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{colnames_l}
\alias{colnames_l}
\title{A lens into the column names of an object}
\format{An object of class \code{lens} of length 2.}
\usage{
colnames_l
}
\description{
The lens version of \code{colnames} and \code{colnames<-}
}
\keyword{datasets}
