% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{dimnames_l}
\alias{dimnames_l}
\title{Dimnames lens}
\format{An object of class \code{lens} of length 2.}
\usage{
dimnames_l
}
\description{
A lens into the dimnames of an object. Lens
equivalent of \link{dimnames} and \link{dimnames<-}.
}
\examples{
x <- matrix(1:4, ncol = 2)
colnames(x) <- c("first", "second")
x

view(x, dimnames_l)
set(x, dimnames_l, list(NULL, c("premiere", "deuxieme")))
}
\keyword{datasets}
