% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-lenses.R
\name{fext_l}
\alias{fext_l}
\title{Create a lens into the file extension}
\usage{
fext_l(extension = c("last", "first"))
}
\arguments{
\item{extension}{last or first indicating what type of extensions to
focus on.}
}
\description{
A lens that focuses on the extension of a filepath
The function creates one of two lenses, \code{last} extension which considers everything
after the last dot (\code{.}) an extension, and \code{first} which considers
everything after the first dot in the basename of the file to be
part of the file extension.
}
\examples{
view("hi.tar.gz", fext_l())
view("hi.tar.gz", fext_l("first"))
set("hi.tar.gz", fext_l(), "xz")
set("hi.tar.gz", fext_l("first"),  "tgz")
set(c("hi.tar.gz", "woah.txt.zip.gz"), fext_l("first"), c("tgz", "wut"))
set(c("hi.tar.gz", "woah.txt.zip.gz"), fext_l(), c("tgz", "wut"))
}
