% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{filter_l}
\alias{filter_l}
\title{Filter lens}
\usage{
filter_l(...)
}
\arguments{
\item{...}{unquoted NSE filter arguments}
}
\description{
Create a lens into the result of a filter. Arguments
are interpreted with non-standard evaluation as in
\link[dplyr:filter]{dplyr::filter}
}
