% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{id_l}
\alias{id_l}
\title{The identity (trivial lens)}
\format{An object of class \code{lens} of length 2.}
\usage{
id_l
}
\description{
This lens focuses on the whole object
}
\examples{
x <- 1:10
view(x, id_l)
head(set(x, id_l, iris))
}
\keyword{datasets}
