% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{lower_tri_l}
\alias{lower_tri_l}
\title{Lens into lower diagonal elements}
\usage{
lower_tri_l(diag = FALSE)
}
\arguments{
\item{diag}{whether or not to include the diagonal}
}
\description{
Create a lens into the lower diagonal elements
of a matrix
}
