% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{over}
\alias{over}
\title{Modify data with a lens}
\usage{
over(d, l, x)
}
\arguments{
\item{d}{the data}

\item{l}{the lens}

\item{x}{the replacement value}
}
\description{
Set the subcomponent of the data referred to by a lens
with a new value. See \link{mkLens} for details.  Merely dispatches
to the \code{over} component of the lens.
}
