% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{rev_l}
\alias{rev_l}
\title{Reverse lens}
\format{An object of class \code{lens} of length 2.}
\usage{
rev_l
}
\description{
Lens into the \link{rev}erse of an object.
}
\keyword{datasets}
