% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\docType{data}
\name{rownames_l}
\alias{rownames_l}
\title{A lens into the row names of an object}
\format{An object of class \code{lens} of length 2.}
\usage{
rownames_l
}
\description{
The lens version of \code{rownames} and \code{rownames<-}
}
\keyword{datasets}
