% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{send_over}
\alias{send_over}
\title{Set one lens to the view of another (transformed)}
\usage{
send_over(d, f, l, m)
}
\arguments{
\item{d}{the data}

\item{f}{the function to apply to the viewed data}

\item{l}{the lens to view through}

\item{m}{the lens to set into}
}
\description{
Set one lens to the view of another (transformed)
}
