% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{slot_l}
\alias{slot_l}
\title{Slot lens}
\usage{
slot_l(slot)
}
\arguments{
\item{slot}{the name of the slot}
}
\description{
The lens equivalent of \code{@} and \code{@<-}
for getting and setting S4 object slots.
}
