% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-lenses.R
\name{strsplit_l}
\alias{strsplit_l}
\title{Strsplit lens}
\usage{
strsplit_l(split)
}
\arguments{
\item{split}{The string to split at, this cannot be a regex (yet?).}
}
\description{
A lens into a split representation of a string. Here
\link{view} is equivalent to \link{strsplit} with \code{fixed = TRUE}
\link{set} is equivalent to \link{paste0} \code{collapse = split} mapped
over each input with \link{vapply}.
}
