% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{upper_tri_l}
\alias{upper_tri_l}
\title{Lens into upper diagonal elements}
\usage{
upper_tri_l(diag = FALSE)
}
\arguments{
\item{diag}{whether or not to include the diagonal}
}
\description{
Create a lens into the upper diagonal elements
of a matrix
}
